fetch(3): Fix wrong usage of proxy when request is redirected to different URL
Under some proxy settings fetch(3) makes wrong usage of proxy when
request is redirected to different URL. (See "2. Detail of the
problem" of bug 252123 for the situation that the problem happens.)
The source of the problem is that in current implementation proxy
usage is decided with the URL of initial request and reused after the
request is redirected to different one. So the problem is fixed by
changing implementation so proxy usage is re-evaluated with new URL
each time request is redirected.