fetch(3); Add support of HTTPS_PROXY/https_proxy environment variables
Add support of HTTPS_PROXY/https_proxy environment variables. They
works similar to FTP_PROXY/ftp_proxy environment variables for FTP
scheme. That is, if scheme of request URL is "https" fetch(3) checks
if following environment variables are set in this order.

1. HTTPS_PROXY
2. https_proxy
3. HTTP_PROXY
4. http_proxy

And the value of first one set is used as setting of proxy for https.